+++ andes/system/__init__.py
@@ -8,6 +8,9 @@ from andes.system.facade import (  # noqa: F401
     example,
     fix_view_arrays,
     import_pycode,
+    load_config_rc,
+    _config_numpy,
 )
 
-__all__ = ["System", "ExistingModels", "example", "fix_view_arrays", "import_pycode"]
+__all__ = ["System", "ExistingModels", "example", "fix_view_arrays",
+           "import_pycode", "_config_numpy", "load_config_rc"]